Extract container metadata from host payload
Container metadata now live in the container util package. We also
remove the registration mechanism and global variable which were a lot
of code to hold 3 callbacks
